requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-versatileimagefield==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-flex-fields==0.8.9
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.1.2
psycopg2==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
python-magic==0.4.22
pytz==2021.1
requests==2.25.1
routers==0.10.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
